[IMP] sale_timesheet: expense in project plan dashboard
On project overview, user can now see the expense
costs of the analytic account linked to the project.
It also displays the reinvoice amount of those costs
(based on their related SO lines), and computed like
the amount to invoice for invoice.

This commit also provides some tests to avoid problem
when modifying the SQL query of profitability report.

Task #49214