[FIX] barcodes: keep vertical position while scanning
Before this commit, you were automatically redirected to the top of
the page when you scanned a product.
It was, therefore, difficult to find the product that had just been
scanned and to check its quantity.

Because of @e0d3311a6 we need to put the focus in an invisible input
at the top of the page.

This input is now always vertically centered in the middle of the page.